scraper/pyproject.toml
[project]
name = "scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "crawl4ai>=0.8.0",
    "markdownify>=1.2.2",
    "python-docx>=1.2.0",
    "requests>=2.32.5",
]
